requirements.txt
Django==3.2.8
django-htmx==1.20.0
djangorestframework==3.12.4
